requirements.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.4.4
psycopg2-binary==2.9.7
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-taggit==6.1.0